@@ -1,15 +1,8 @@
-import Vue from 'vue'
 import Vuex from 'vuex'
 
-Vue.use(Vuex)
-
-export default new Vuex.Store({
-  state: {
-  },
-  mutations: {
-  },
-  actions: {
-  },
-  modules: {
-  }
+export default Vuex.createStore({
+  state: {},
+  mutations: {},
+  actions: {},
+  modules: {}
 })
